pyproject.toml
[project]
name = "test-hea-llama"
version = "6.0.0"
description = "Autonomous, self-learning AI test generator for Java/Spring Boot"
authors = [{ name = "itbuddy" }]
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.4",
    "langchain-ollama>=0.2.0",
    "chromadb>=0.5.23",
    "beautifulsoup4>=4.14.3",
    "requests>=2.32.5",
    "click>=8.3.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
    "rank_bm25>=0.2.2",
    "flashrank>=0.2.10",
    "duckduckgo-search>=8.1.1",
    "mcp>=1.25.0",
    "langchain-mcp-adapters>=0.2.1",
    "pytest",
    "pytest-mock",
    "pyyaml"
]

[project.scripts]
llama = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
